app/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.7
names==0.3.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
trueskill==0.4.4
wcwidth==0.1.7
